+++ app/Filament/Resources/Members/Pages/ListMembers.php
@@ -3,7 +3,6 @@
 namespace App\Filament\Resources\Members\Pages;
 
 use App\Filament\Resources\Members\MemberResource;
-use Filament\Actions\CreateAction;
 use Filament\Resources\Pages\ListRecords;
 
 class ListMembers extends ListRecords
@@ -12,8 +11,6 @@ class ListMembers extends ListRecords
 
     protected function getHeaderActions(): array
     {
-        return [
-            CreateAction::make(),
-        ];
+        return [];
     }
 }