+++ resources/views/filament/membership-form-link.blade.php
@@ -1,0 +1,10 @@
+<x-filament::button
+    class="ditib-brand-primary-button"
+    color="primary"
+    href="https://mitglied.ditib-ahlen-projekte.de/"
+    icon="heroicon-m-user-plus"
+    tag="a"
+    target="_blank"
+>
+    Neue Registrierung
+</x-filament::button>